test/CMakeLists.txt: Tolerate out-of-tree builds of OpenSSL
Add a new variable `OEPNSSL_SRCTOP' to parallel the existing
`OPENSSL_BLDTOP'.  This refers to the actual source directory, from
which we find header files, rather than the build directory which
contains libraries, most notably `libtestutil.a'.

As I've set things up, each variable defaults from the other.  This is
probably not great long-term: it's intended as a transition measure.  It
seems more sensible to me to have the source directory be the primary
setting, and have the build directory be subsidiary; but existing
scripts (and finger-macros) know nothing of the new source-directory
setting, so preserving compatibility seems valuable for now.

Signed-off-by: Mark Wooding <mark.wooding@trustonic.com>