Fix generate.py duplicate hybrid NID warning
It appears that the erroneous warning message motivating this commit is
due to some vestigial code special-casing x-curves. The
`nid_ecx_hyybrid` kem field was only used to generate the duplicate ecx
nid warning message, so we track duplicates in a local `set` instead.
Additionally, we expand the warning message to all curve NIDs, not just
x-curves and treat duplicates as fatal because extra curve NIDs should
be unique for each KEM.