Build a module instead of a shared library.
Fix #202.

This commit switches the default type in CMake from `SHARED` to `MODULE`.

Two issues have been found.

## First issue

I ran into an issue while testing this commit on macOS. Here is the error I
got from clang:

```shell
clang: error: invalid argument '-compatibility_version 1.0.0' only allowed with '-dynamiclib'
```

After some investigation, I found out that the [`SOVERSION` and `VERSION`](https://cmake.org/cmake/help/latest/prop_tgt/SOVERSION.html)
properties should not be used when building a shared library. This doesn't seem
to lead to an error on Linux, but on macOS, it happens to set the following
two flags:
  * `-compatibility_version` (given by `SOVERSION`)
  * `-current_version` (given by `VERSION`)

However, CMake uses the `-bundle` flag when linking, instead of the `-dynamiclib`
flag.

I found a fix by reading the documentation:

> For shared libraries, the MACHO_COMPATIBILITY_VERSION and MACHO_CURRENT_VERSION
> properties can be used to override the compatibility version and current
> version respectively.

Setting these two flags to `OFF` removes `-compatibility_version` and
`-current_version`.

Perhaps we should simply remove `SOVERSION` and `VERSION`, as they're only
meant to be used for shared libraries?

## Second issue

OpenSSL has an API to load dynamic objects and libraries called `DSO`.
When loading a provider, a call to `dlfcn_name_converter` is made to compute
the name of the file:

https://github.com/openssl/openssl/blob/79fa5873a9c2f2df392bc4b39146d8db37640118/crypto/dso/dso_dlfcn.c#L253-L280

Reader may notice that OpenSSL appends the macro `DSO_EXTENSION` to the end
of the provider's name.
`DSO_EXTENSION` is platform-specific: its value is `.so` on Unix, and `.dylib`
on macOS.
Because we're building a `MODULE` library here instead of a `SHARED` library,
the output file's extension is `.so` instead of `.dylib`. Thus, OpenSSL fails
to find the `oqs-provider` to load.

To solve this issue, I used the [`SUFFIX`](https://cmake.org/cmake/help/latest/prop_tgt/SUFFIX.html)
property of CMake targets to explicitely specify the extension when building for
a XNU-based platform (macOS, iOS, watchOS, etc.).


also change suffix on Windows.

adapt CI to MODULE-defined paths.