Fix a minor bug in the `runtests.sh`.
This commit fixes a minor bug in `scripts/runtests.sh`: this script is run with
POSIX sh on linux machines, meaning that `[[` are not recognized.
However, `[[ … ]]` was used (see for instance [this CI job]).

Because `set -e` was not set, the minor bug described above did not lead to a
CI failure.

This commit fixes the `[[ … ]]` by using a simple `if grep`.
It also sets `set -e` at the top of the file in order to catch any potential future mistakes.

Finally, this commit also fixes minor warnings raised by [ShellCheck]. To display those warnings,
one can simply run `shellcheck scripts/runtests.sh`.

[this CI job]: https://github.com/open-quantum-safe/oqs-provider/actions/runs/5527705776/jobs/10083733262#step:6:14
[ShellCheck]: https://www.shellcheck.net/