Add a CircleCI job to test oqs-provider against memory leaks.
This commit introduces a new CircleCI job called `check-ASan-tests` that compiles
OpenSSL 3, liboqs and oqs-provider with ASan (`-fsanitize=address`). Then, it
runs the CTest suite.

If any memory leak occurs somewhere and is triggered in one of the various unit tests,
then this CircleCI job fails and displays the ASan trace.