Fix #338 and #339: output a valid aarch64 debian package with a valid directory layout.
This commit aims to fix #338 and #339.

It fixes #338 by setting [`CPACK_DEBIAN_PACKAGE_ARCHITECTURE`] to the architecture
used by CMake. To do so, we inspect the value of [`CMAKE_SYSTEM_PROCESSOR`], which
tells us the target architecture.

A test has been added to ensure that the target architecture of the output .deb
is `arm64`.


This commit also fixes #339 by creating a configuration called `DebPack` which
sets the right file paths for `cpack`.

When `cpack` is invoked with `-C DebPack`, `lib/ossl-modules`, `lib` and `include/oqs-provider`
are being used as destination directories.

[`CPACK_DEBIAN_PACKAGE_ARCHITECTURE`]: https://cmake.org/cmake/help/latest/cpack_gen/deb.html#variable:CPACK_DEBIAN_PACKAGE_ARCHITECTURE
[`CMAKE_SYSTEM_PROCESSOR`]: https://cmake.org/cmake/help/latest/variable/CMAKE_SYSTEM_PROCESSOR.html#variable:CMAKE_SYSTEM_PROCESSOR

Signed-off-by: thb-sb <is@md5secure.com>