Hide all symbols except for OSSL_provider_init entrypoint
When oqsprovider is loaded into an application which already defines
symbols such as `setECParams` or `oqs_prov_bio_from_dispatch`, crashes
can occur. Therefore restrict the visibility to hidden by default, this
is similar to use of the `OQS_API` macro in liboqs.

Similar issues can occur when a statically linked liboqs is loaded into
an application that already loaded liboqs dynamically.

With the above fixes, the risk of conflicts and crashes on Linux is
reduced as `OSSL_provider_init` is the only exposed symbol.

Signed-off-by: Peter Wu <pwu@cloudflare.com>